IoT/requirements.txt
ultralytics
opencv-python
mysql-connector